int, long => int32

+++ include/audio_stream.h
@@ -29,7 +29,7 @@ class AudioStream
 {
 public:
 	virtual ~AudioStream() {};
-	virtual long read(void *buffer, size_t frame_count) = 0;
+	virtual int32_t read(void *buffer, size_t frame_count) = 0;
 	virtual bool seek(size_t frame_no) = 0;
 	virtual int32_t frame_rate() const = 0; /* in PCM frames per sec */
 	virtual int channels() const = 0;
